+++ kitchenpos-table/src/main/java/kitchenpos/table/ordertable/ui/TableRestController.java
@@ -1,0 +1,59 @@
+package kitchenpos.table.ordertable.ui;
+
+import kitchenpos.table.ordertable.application.TableService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import kitchenpos.table.ordertable.domain.OrderTable;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class TableRestController {
+    private final TableService tableService;
+
+    public TableRestController(final TableService tableService) {
+        this.tableService = tableService;
+    }
+
+    @PostMapping("/api/tables")
+    public ResponseEntity<OrderTable> create(@RequestBody final OrderTable orderTable) {
+        final OrderTable created = tableService.create(orderTable);
+        final URI uri = URI.create("/api/tables/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created)
+                ;
+    }
+
+    @GetMapping("/api/tables")
+    public ResponseEntity<List<OrderTable>> list() {
+        return ResponseEntity.ok()
+                .body(tableService.list())
+                ;
+    }
+
+    @PutMapping("/api/tables/{orderTableId}/empty")
+    public ResponseEntity<OrderTable> changeEmpty(
+            @PathVariable final Long orderTableId,
+            @RequestBody final OrderTable orderTable
+    ) {
+        return ResponseEntity.ok()
+                .body(tableService.changeEmpty(orderTableId, orderTable))
+                ;
+    }
+
+    @PutMapping("/api/tables/{orderTableId}/number-of-guests")
+    public ResponseEntity<OrderTable> changeNumberOfGuests(
+            @PathVariable final Long orderTableId,
+            @RequestBody final OrderTable orderTable
+    ) {
+        return ResponseEntity.ok()
+                .body(tableService.changeNumberOfGuests(orderTableId, orderTable))
+                ;
+    }
+}
