+++ kitchenpos-table/src/main/java/kitchenpos/table/tablegroup/domain/TableGroup.java
@@ -1,0 +1,38 @@
+package kitchenpos.table.tablegroup.domain;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "table_group")
+public class TableGroup {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(name = "created_date")
+    private LocalDateTime createdDate;
+
+    private TableGroup(LocalDateTime createdDate) {
+        this.createdDate = createdDate;
+    }
+
+    public static TableGroup from(LocalDateTime createdDate) {
+        return new TableGroup(createdDate);
+    }
+
+    protected TableGroup() {
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public LocalDateTime getCreatedDate() {
+        return createdDate;
+    }
+}
