+++ kitchenpos-table/src/main/java/kitchenpos/table/tablegroup/domain/TableGroupRepository.java
@@ -1,0 +1,6 @@
+package kitchenpos.table.tablegroup.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
+}
