+++ kitchenpos-table/src/main/java/kitchenpos/table/tablegroup/dto/CreateTableGroupRequest.java
@@ -1,0 +1,19 @@
+package kitchenpos.table.tablegroup.dto;
+
+import java.util.List;
+
+public class CreateTableGroupRequest {
+
+    private List<Long> orderTableIds;
+
+    public CreateTableGroupRequest(List<Long> orderTableIds) {
+        this.orderTableIds = orderTableIds;
+    }
+
+    protected CreateTableGroupRequest() {
+    }
+
+    public List<Long> getOrderTableIds() {
+        return orderTableIds;
+    }
+}
