+++ kitchenpos-table/src/main/java/kitchenpos/table/tablegroup/event/TableGroupEventPublisher.java
@@ -1,0 +1,19 @@
+package kitchenpos.table.tablegroup.event;
+
+import org.springframework.context.ApplicationEvent;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.stereotype.Component;
+
+@Component
+public class TableGroupEventPublisher {
+
+    private final ApplicationEventPublisher applicationEventPublisher;
+
+    public TableGroupEventPublisher(ApplicationEventPublisher applicationEventPublisher) {
+        this.applicationEventPublisher = applicationEventPublisher;
+    }
+
+    public void publish(ApplicationEvent applicationEvent) {
+        applicationEventPublisher.publishEvent(applicationEvent);
+    }
+}
