+++ kitchenpos-table/src/main/java/kitchenpos/table/tablegroup/ui/TableGroupRestController.java
@@ -1,0 +1,39 @@
+package kitchenpos.table.tablegroup.ui;
+
+import kitchenpos.table.tablegroup.application.TableGroupService;
+import kitchenpos.table.tablegroup.domain.TableGroup;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import kitchenpos.table.tablegroup.dto.CreateTableGroupRequest;
+
+import java.net.URI;
+
+@RestController
+public class TableGroupRestController {
+    private final TableGroupService tableGroupService;
+
+    public TableGroupRestController(final TableGroupService tableGroupService) {
+        this.tableGroupService = tableGroupService;
+    }
+
+    @PostMapping("/api/table-groups")
+    public ResponseEntity<TableGroup> create(@RequestBody final CreateTableGroupRequest request) {
+        final TableGroup created = tableGroupService.create(request);
+        final URI uri = URI.create("/api/table-groups/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created)
+                ;
+    }
+
+    @DeleteMapping("/api/table-groups/{tableGroupId}")
+    public ResponseEntity<Void> ungroup(@PathVariable final Long tableGroupId) {
+        tableGroupService.ungroup(tableGroupId);
+        return ResponseEntity.noContent()
+                .build()
+                ;
+    }
+}
