+++ kitchenpos-table/src/test/java/kitchenpos/TableApplication.java
@@ -1,0 +1,12 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TableApplication {
+
+    @Test
+    public void contextLoads() {
+    }
+}
