--- src/main/java/kitchenpos/tablegroup/event/TableGroupEventPublisher.java
@@ -1,19 +1,0 @@
-package kitchenpos.tablegroup.event;
-
-import org.springframework.context.ApplicationEvent;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Component;
-
-@Component
-public class TableGroupEventPublisher {
-
-    private final ApplicationEventPublisher applicationEventPublisher;
-
-    public TableGroupEventPublisher(ApplicationEventPublisher applicationEventPublisher) {
-        this.applicationEventPublisher = applicationEventPublisher;
-    }
-
-    public void publish(ApplicationEvent applicationEvent) {
-        applicationEventPublisher.publishEvent(applicationEvent);
-    }
-}
